Working on castling

+++ Chess/DDHChess/ChessPlayer.h
@@ -13,7 +13,7 @@
 
 
 // Make our lives easier by defining chess players as ints
-typedef NS_ENUM(NSUInteger, ChessPlayer){
+typedef NS_ENUM(int, ChessPlayer){
     ChessPlayerWhite = 0,
     ChessPlayerBlack = 1,
     ChessPlayerNull = 2
